Use longs for money in the model

+++ app/src/main/java/fr/epf/matmob/filmroll/model/Film.kt
@@ -18,7 +18,7 @@ data class Film(
     val voteScore: Float,
     val voteCount: Int,
     val collection: String,
-    val budget: Int,
+    val budget: Long,
     val genres: List<Genre>,
     val homepage: URL,
     val imdbId: String,
